+++ ccramic/components/layout.py
@@ -17,7 +17,6 @@ import dash_tour_component
 from ccramic.utils.alert import DataImportTour, ToolTips
 from ccramic.io.session import SessionTheme, TabText
 
-
 def register_app_layout(config, cache_dest):
 
     # set the default colours for the swatches from the config input
@@ -232,7 +231,7 @@ def register_app_layout(config, cache_dest):
                                         html.H5("Import quantification results"),
                                         du.Upload(id='upload-quantification', max_file_size=5000,
                                                   filetypes=['h5ad', 'h5', 'csv'],
-                                                  text='Import cell quantification results in CSV format using drag and drop',
+                                                  text='Import quantification results in CSV format using drag and drop',
                                                   max_files=1, upload_id="upload-quantification",
                                                   default_style={"margin-top": "20px", "height": "10%"}),
                                         html.Br(),
@@ -611,7 +610,7 @@ def register_app_layout(config, cache_dest):
                                         html.Div([dcc.Dropdown([1, 3, 5, 7, 9], 3, id='global-kernel-val-filter',
                                                                style={"width": "75%"}),
                                                   dcc.Input(id="global-sigma-val-filter", type="number", value=1,
-                                                            style={"width": "60%"},
+                                                            style={"width": "60%"}, debounce=True,
                                                             disabled=True, min=0, max=9, step=0.1)],
                                                  style={"display": "flex", "margin": "20px"}),
                                         # dbc.Button("Reset all channels to default", id="set-default-rangeslider-all",
@@ -637,13 +636,12 @@ def register_app_layout(config, cache_dest):
                                                                              labelPosition='bottom',
                                                                              color=DEFAULT_WIDGET_COLOUR,
                                                                              style={"margin-left": "60px"}),
-                                                            html.Abbr(dcc.Checklist(options=[' add boundary'],
+                                                            dcc.Checklist(options=[' add boundary'],
                                                                                     value=[' add boundary'],
                                                                                     id="add-mask-boundary",
                                                                                     style={"margin-left": "35px",
                                                                                            "margin-top": "10px",
                                                                                            "accent-color": DEFAULT_WIDGET_COLOUR}),
-                                                                      title="Use this feature only if the cell boundary was not derived on import"),
                                                             dcc.Checklist(options=[' Show mask ID on hover'],
                                                                           value=[], id="add-cell-id-mask-hover",
                                                                           style={"margin-left": "35px",
@@ -686,7 +684,7 @@ def register_app_layout(config, cache_dest):
                                                 html.H6(children=[], id="gating-param-display", style={"width": "60%"}),
                                                 ], style={"display": "flex", "width": "100%", "justifyContent": "center"}),
                                             html.Br(),
-                                            html.H6("Annotate gated cells", style={"width": "75%"}),
+                                            html.H6("Annotate gated objects", style={"width": "75%"}),
                                             html.Div([dcc.Dropdown(id='quant-annotation-col-gating',
                                             multi=False, options=['ccramic_cell_annotation'],
                                             value="ccramic_cell_annotation", style={"width": "80%"}),
@@ -781,7 +779,7 @@ def register_app_layout(config, cache_dest):
                                                 id="click-annotation-add-circle", style={"margin-top": "12px",
                                                 "accent-color": DEFAULT_WIDGET_COLOUR}),
                                 dcc.Input(id="click-annotation-assignment", type="text", value=None,
-                                          placeholder="Add a cell type for click",
+                                          placeholder="Add an annotation on click",
                                     style={"width": "65%", "margin-left": "1px"})],
                                          style={"display": "flex"}),
                                 dcc.Checklist(options=[' Overlay grid'], value=[],
@@ -847,7 +845,7 @@ def register_app_layout(config, cache_dest):
                                 dcc.Download(id="download-region-csv"),
                                 dbc.Modal(children=dbc.ModalBody(
                                 [dbc.Row([dbc.Col([html.H6("Create a region annotation")], width=8),
-                                          dbc.Col([html.H6("Annotate with cell type")], width=4)]),
+                                          dbc.Col([html.H6("Add object annotation")], width=4)]),
                                  dbc.Row([dbc.Col([html.Div([dcc.Input(id="new-annotation-col", type="text",
                                 value="", placeholder="Create annotation column",
                                 style={"width": "50%", "margin-right": "10px", "height": "50%"}),
@@ -867,10 +865,8 @@ def register_app_layout(config, cache_dest):
                                           style={"width": "65%", "margin-right": "10px", "height": "50%"})],
                                 style={"display": "flex"})], width=8),
                                 dbc.Col([
-                                # dcc.Dropdown(id='region-annotation-cell-types',
-                                # multi=False, options=[], placeholder="Select a cell type")
                                 dcc.Input(id="region-annotation-cell-types", type="text",
-                                value="", placeholder="New cell type", style={"width": "65%", "margin-right": "10px",
+                                value="", placeholder="New annotation type", style={"width": "65%", "margin-right": "10px",
                                     "height": "100%"})
                                 ], width=4)]),
                                 dbc.Row(dbc.Col(html.Div([], style={"display": "flex"}))),
@@ -980,7 +976,7 @@ def register_app_layout(config, cache_dest):
 
             dbc.Tab(label="Channel gallery", tab_id='gallery-tab', id='gallery-tab',
                     label_style={"color": DEFAULT_WIDGET_COLOUR},
-                        children=[
+                        children=[dcc.Loading([
                             html.Br(),
                             html.H6(TabText().channel_tiles),
                             html.Div([daq.ToggleSwitch(label='Change thumbnail on zoom',
@@ -1004,7 +1000,7 @@ def register_app_layout(config, cache_dest):
                                            style={"display": "flex"}),
                         html.Div(id="image-gallery", children=[
                         dbc.Row(id="image-gallery-row")], style={"margin-top": "15px"}),
-                                  ]),
+                        ], type="default", fullscreen=True, color=DEFAULT_WIDGET_COLOUR)]),
             dbc.Tab(label="Panel metadata", tab_id='metadata-tab', label_style={"color": DEFAULT_WIDGET_COLOUR},
                     children=
                         [html.Div([dbc.Row([
@@ -1027,7 +1023,7 @@ def register_app_layout(config, cache_dest):
                                   children=[
                                 html.Div([dbc.Row([
                                 dbc.Col(html.Div([html.Br(),
-                                html.Div([html.B("Cell-level marker expression", style={"margin-bottom": "10px"}),
+                                html.Div([html.B("Object-level marker expression", style={"margin-bottom": "10px"}),
                                 daq.ToggleSwitch(label='Normalize heatmap', id='normalize-heatmap', labelPosition='bottom',
                                     color=DEFAULT_WIDGET_COLOUR, value=True),
                                 dcc.Input(type="number", placeholder="Subset heatmap", id="subset-heatmap",
@@ -1091,7 +1087,7 @@ def register_app_layout(config, cache_dest):
                                         id="dynamic-update-barplot", style={"margin-top": "12px",
                                             "accent-color": DEFAULT_WIDGET_COLOUR}),
                                         html.Br(),
-                                        html.H6("Add annotation to cells"),
+                                        html.H6("Add annotation to UMAP"),
                                         dcc.Input(id="annotation-col-quantification", type="text",
                                         value="", placeholder="New annotation category",
                                         style={"width": "55%", "margin-right": "10px", "height": "100%",
@@ -1103,7 +1099,7 @@ def register_app_layout(config, cache_dest):
                                         multi=False, options=['ccramic_cell_annotation'],
                                             value="ccramic_cell_annotation", style={"width": "75%"}),
                                         dcc.Input(id="annotation-cell-types-quantification", type="text",
-                                        value="", placeholder="New cell type",
+                                        value="", placeholder="New annotation type",
                                         style={"width": "55%", "margin-right": "10px", "height": "100%",
                                                "margin-top": "5px"}),
                                         dbc.Button("Annotate UMAP", id="create-annotation-umap",
@@ -1144,7 +1140,7 @@ def register_app_layout(config, cache_dest):
                                     html.Br(),
                                     html.Br()
                                       ])]),
-                        dbc.Modal(children=dbc.ModalBody([html.H6("Select the cell type annotation column"),
+                        dbc.Modal(children=dbc.ModalBody([html.H6("Select the object type annotation column"),
                         dcc.Dropdown(id='cell-type-col-designation',
                             multi=False, options=[], style={'width': '100%'})]),
                         id="quantification-config-modal", size='l', style={"margin-left": "10px",