+++ ccramic/utils/pixel_level_utils.py
@@ -232,8 +232,11 @@ def pixel_hist_from_array(array, subset_number=1000000, keep_max=True):
                                    np.array([max_hist]).astype(cast_type)])
     except ValueError:
         pass
-    return go.Figure(px.histogram(hist, range_x=[min(hist), max_hist]), layout_xaxis_range=[0, max_hist]), \
-        float(np.max(array))
+
+    fig = go.Figure(px.histogram(hist, range_x=[min(hist), max_hist]), layout_xaxis_range=[0, max_hist])
+    fig.update_layout(showlegend=False, yaxis={'title': None},
+                                      xaxis={'title': None}, margin=dict(pad=0))
+    return fig, float(np.max(array))
 
 def upper_bound_for_range_slider(array):
     """